jdt: eagerly detect and report non-incremental annotation processors
compiler requires one and only one originating element in order to be
able to cleanup generated resources during incremental build. annotation
processors that do not provide such original element can only be used
reliably with proc set to 'procEX' or 'onlyEX' and result in immediate
build failure if used with proc set to 'proc' or 'only'.

Signed-off-by: Igor Fedorenko <ifedorenko@salesforce.com>