introduced compiler dependencySourceTypes=ignore|prefer parameter
classpath lookup can find .class, .java or both. when both are found,
javac used to choose the file with the latest filesystem timestamp,
which resulted in semi-random and hard to troubleshoot errors
depending which file it chose. jdt preferred .class files.

the new parameter lets choose if the compiler should ignore .java files
(the default) or prefer them.

Signed-off-by: Igor Fedorenko <ifedorenko@salesforce.com>