Accommodate an alternate lifecycle producing the primary artifact. (#74)
Currently during the JAR producing Takari lifecycles (takari-jar, takari-maven-component and takari-maven-plugin) make the reasonable assumption that the Jar Mojo will be producing the primary artifact. This change allows a flag to be set during the build to signal that another Takari-based lifecycle will produce the primary artifact.

In my particular use case I have a "provisio" lifecycle that uses Takari Mojos for standard lifecycle handling but it produces a tar.gz file of a server runtime as the primary artifact. I want the JAR to be an attached artifact and packaged with the server but also deployed separately.

This change should not interfere with any existing cases as in order to trigger this behavior the user has to explicity use a custom lifecycle, and explicity set the flag to take over to role of producing the primary artifact.