Support parallel deploy
Maven Resolver 1.9.5 new feature is ability of parallel deploy,
and to be utilize it, ideally one should "pack up" all into
single DeployRequest and let resolver handle it.

Hence, this change has:
* cosmetic name changes in Deploy mojo
* batches DeployRequests, as currently the are per-module. It would
work like it still, but the ideal is to have all in there and let
resolver just figure it out.

The parallel deploy is transparently available with resolver 1.9.5,
(maven 3.9.1), so it will transparently become available to
takari lifecycle as well, if used with Maven 3.9.1. On older
version this has no any (side) effect, just that the artifacts
and metadata will be deployed in-order: first ALL artifacts, then
ALL metadata (in proper order).